tests: move util tests to __tests__ dir.
